requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.0.3
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
chromadb==0.4.18
sentence-transformers==2.2.2
langchain==0.0.340
spacy==3.7.2
scikit-learn==1.3.2
textstat==0.7.3
celery==5.3.4
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
alembic==1.12.1
httpx==0.25.2